Clean up newCMAESInstance #668
Moved newCMAESInstance out of improvement and optimizing emitter since it was just being repeated
