+++ src/main/java/edu/southwestern/evolution/mapelites/emitters/Emitter.java
@@ -3,9 +3,12 @@ package edu.southwestern.evolution.mapelites.emitters;
 import java.util.ArrayList;
 
 import edu.southwestern.MMNEAT.MMNEAT;
+import edu.southwestern.evolution.genotypes.Genotype;
 import edu.southwestern.evolution.mapelites.Archive;
 import edu.southwestern.evolution.mapelites.CMAME;
 import edu.southwestern.log.MMNEATLog;
+import edu.southwestern.parameters.Parameters;
+import edu.southwestern.util.datastructures.ArrayUtil;
 import fr.inria.optimization.cmaes.CMAEvolutionStrategy;
 
 public abstract class Emitter implements Comparable<Emitter> {
@@ -60,8 +63,24 @@ public abstract class Emitter implements Comparable<Emitter> {
 	 * @param archive
 	 * @return
 	 */
-	protected abstract CMAEvolutionStrategy newCMAESInstance(Archive<ArrayList<Double>> archive); 
-	
+	protected CMAEvolutionStrategy newCMAESInstance(Archive<ArrayList<Double>> archive) {
+		CMAEvolutionStrategy optEmitter = new CMAEvolutionStrategy();
+		optEmitter.setDimension(dimension);
+		Genotype<ArrayList<Double>> elite = archive.getElite(archive.randomOccupiedBinIndex()).individual;
+		double[] phenod = ArrayUtil.doubleArrayFromList(elite.getPhenotype());
+		optEmitter.setInitialX(phenod); // start at random bin
+		optEmitter.setInitialStandardDeviation(0.5); // unsure if should be hardcoded or not
+		int lambda = Parameters.parameters.integerParameter("lambda"); 
+		// Realized that mu = lambda / 2 after extensively reviewing the pyribs code and walking through with Amy Hoover.
+		// Even in pyribs, there is apparently an option for mu to be set differently, but it is definitely fixed, and
+		// the default is for it to be half of lambda.
+		int mu = lambda/2; 
+		optEmitter.parameters.setMu(mu);
+		optEmitter.parameters.setPopulationSize(lambda);
+		optEmitter.init();
+		optEmitter.writeToDefaultFilesHeaders(0); // Overwrite existing CMA-ES files
+		return optEmitter;
+	}
 	
 	/**
 	 * Update the internal CMA-ES instance distribution
